Use streaming when creating log symbols file.
